tools: kwbimage: do not build security features without LIBCRYPTO
Commit <b4f3cc2c> "Tools: kwbimage: Do not hide usage of secure header under
CONFIG_ARMADA_38X" allowed the build of secure header features for the
host tool unconditionally. This makes sens, as the host tool should be as
feature-complete as possible, whether or not the target supports the security
features itself. But this breaks the compilation of host tool if
"CONFIG_TOOLS_LIBCRYPTO" is not enabled.

Reverts a part of <b4f3cc2c>, but use "CONFIG_TOOLS_LIBCRYPTO" dependency instead
of deprecated "CONFIG_KWB_SECURE" used before.